Chain of responsability pattern
Implementiran chain of responsability pattern, kao i deo aplikacije koji ne koristi pomenuti pattern.

+++ DesignPatterns/pizzaOrder/pattern/pizza's/AbstractPizza.cs
@@ -10,7 +10,7 @@ namespace DesignPatterns.pizzaOrder.pattern
 
         protected String name = "No name";
 
-        public abstract double cost();
+        public abstract int cost();
         
         public String getName()
         {
